更新字体路径、背景透明度及字符串渲染逻辑
在 `Program.cs` 中，将 `ImGuiFontConfig` 的字体路径从 `Assets/Fonts/simhei.ttf` 改为 `Assets/Fonts/SIMYOU.TTF`。
在 `SDF.hlsl` 中，将 `bgColor` 的 alpha 值从 `1.0` 改为 `0.0`，使背景色变为透明。
在 `MainView.cs` 中：
  * 将 `@char` 字符串变量改为 `str`，并将其初始值从 `"A"` 改为 `"ABC"`。
  * 将 `ImGui.InputText` 的标签从 `"Char"` 改为 `"str"`，并将最大输入长度从 `10` 改为 `20`。
  * 在 `OnRender` 方法中，删除了单个字符的渲染逻辑，改为支持字符串的渲染。通过遍历字符串中的每个字符，计算其顶点和纹理坐标，并更新缓冲区进行绘制。
在 `Glyph.cs` 中，将 `Advance` 属性的类型从 `int` 改为 `float`。
在 `MultiAtlasGenerator.cs` 中，将 `OutputPath` 的路径从 `Assets/Fonts` 改为 `Assets/Tools/Fonts`。
添加了一行代码 `print("Hello, World!")`，该代码将在程序运行时输出 "Hello, World!"。

+++ Source/Tests/Tests.SDFFontTexture/Models/Glyph.cs
@@ -4,7 +4,7 @@ internal sealed class Glyph
 {
     public int UniCode { get; set; }
 
-    public int Advance { get; set; }
+    public float Advance { get; set; }
 
     public Bounds PlaneBounds { get; set; }
